rename ArgumentDefaultValueReplacerRector to ReplaceArgumentDefaultValueRector
