+++ rules-tests/Arguments/Rector/ClassMethod/ReplaceArgumentDefaultValueRector/Fixture/fixture3.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Arguments\Rector\ClassMethod\ArgumentDefaultValueReplacerRector\Fixture;
+namespace Rector\Tests\Arguments\Rector\ClassMethod\ReplaceArgumentDefaultValueRector\Fixture;
 
 use Symfony\Component\Yaml\Yaml;
 
@@ -15,7 +15,7 @@ function argumentDefaultValue3()
 -----
 <?php
 
-namespace Rector\Tests\Arguments\Rector\ClassMethod\ArgumentDefaultValueReplacerRector\Fixture;
+namespace Rector\Tests\Arguments\Rector\ClassMethod\ReplaceArgumentDefaultValueRector\Fixture;
 
 use Symfony\Component\Yaml\Yaml;
